refactor(tests): mover plan-related tests a tests/plan/ (issue #368)
Primera oleada de issue #368 (reorganización tests/). El directorio
tests/ tiene 264 archivos en root sin estructura clara, dificultando
distinguir unit vs integration vs E2E.

Esta PR mueve 14 archivos relacionados con kj plan a tests/plan/:
- plan-auto-certify-anomaly, plan-module, plan-fixer, plan-reviewer
- planner-role, tests-synthesizer, plan-hu-outcome
- plan-ready-tests-first-gate, plan-generate-no-chain
- plan-store-kj-home-isolation, plan-to-run
- planning-game-adapter, plan-spec-mapping, plan-tests-first

Cambios mecánicos:
- git mv preserva history (rename)
- sed corrige imports relativos: ../src → ../../src
- sed corrige vi.mock paths

Tests: 14/14 archivos verde (157 tests), suite global 4577/4577.

Próximas oleadas (PRs separadas):
- tests/hu/ (hu-decomposition, hu-git-automation, hu-board-sync, ...)
- tests/orchestrator/ (varios ya tocados)
- tests/sonar/ etc.
